+++ src/error.rs
@@ -1,5 +1,3 @@
-use ring::error::KeyRejected;
-use ring::error::Unspecified;
 use std::{error::Error as Err, fmt};
 
 #[derive(Debug)]
@@ -13,17 +11,14 @@ pub enum Error {
     Json(serde_json::Error),
     Auth(AuthError),
     #[cfg(feature = "jwt")]
-    InvalidRsaKey(Unspecified),
+    InvalidRsaKey(ring::error::Unspecified),
     #[cfg(feature = "jwt")]
-    InvalidRsaKeyRejected(KeyRejected),
+    InvalidRsaKeyRejected(ring::error::KeyRejected),
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use Error::{
-            Auth, Base64Decode, Http, HttpStatus, InvalidKeyFormat, InvalidRsaKey,
-            InvalidRsaKeyRejected, Io, Json,
-        };
+        use Error::*;
 
         match self {
             Io(err) => write!(f, "{}", err),