+++ src/gcp/jwt.rs
@@ -114,7 +114,7 @@ impl Default for Algorithm {
     }
 }
 
-/// The supported RSA key formats, see the documentation for `ring::signature::RsaKeyPair`
+/// The supported RSA key formats, see the documentation for [`ring::signature::RsaKeyPair`]
 /// for more information
 pub enum Key<'a> {
     /// An unencrypted PKCS#8-encoded key. Can be used with both ECDSA and RSA