+++ src/token.rs
@@ -34,7 +34,7 @@ impl Token {
         self.access_token.is_empty() || self.expiry_date() <= Utc::now()
     }
 
-    /// Returns a `DateTime` object representing our expiry date.
+    /// Returns a [`chrono::DateTime`] object representing our expiry date.
     pub fn expiry_date(&self) -> DateTime<Utc> {
         match self.expires_in_timestamp {
             Some(ts) => Utc.timestamp(ts, 0),