+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added new field to `Error::InvalidRsaKey`
 - Added new Error varient `Error::InvalidRsaKeyRejected` 
+
 ## [0.4.7] - 2021-01-18
 ### Changed
 - Updated `base64` to `0.13`, matching the version used by rustls